+++ discord/eurobot.v13/lib/commands/help.ts
