+++ discord/eurobot.v13/lib/commands/ping.ts
