+++ discord/eurobot.v13/lib/events/ready.ts
