+++ discord/eurobot.v13/lib/models/discord.ts
