+++ discord/eurobot.v13/lib/models/news.ts
