+++ discord/eurobot.v13/lib/tools.ts
