+++ discord/eurobot.v13/tslint.json
