+++ discord/eurobot.v13/types/roll.d.ts
