+++ discord/eurobot/package.json
