+++ discord/eurobot/sql/20210727.sql
