+++ discord/eurobot/src/eurobot.ts
