+++ discord/eurobot/src/lib/tools.ts
