+++ discord/eurobot/src/models/discord-calendar.ts
