+++ discord/eurobot/src/models/discord-message.ts
