+++ discord/eurobot/src/models/discord-news.ts
