+++ discord/eurobot/src/models/discord-poll.ts
