+++ discord/eurobot/src/models/discord-user.ts
