+++ discord/eurobot/src/types/index.d.ts
