+++ discord/eurobot/src/types/roll.d.ts
