+++ discord/eurobot/tsconfig.json
