+++ discord/eurobot/tslint.json
