--- src/discord/eurobot/src/conf/discord.json
@@ -1,5 +1,0 @@
-{
-
-    "maxUserWarnings":3
-
-}
