--- src/discord/eurobot/src/conf/google/options.json
@@ -1,8 +1,0 @@
-{
-    "calendar":  {
-        "calendarID":"6k0f97k0bbrl7ub7es5iqijldc@group.calendar.google.com",
-        "calendarLocation":"https://calendar.google.com/calendar/embed?src=6k0f97k0bbrl7ub7es5iqijldc%40group.calendar.google.com",
-        "defaultCalendarChannel":"",
-        "maxItems":5
-    }
-}
